backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
Pillow>=10.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0